Rename of RpcRecord -> RpcResult and moved code to RpcResult.cc/.h
