+++ src/AbstractLog.cc
@@ -200,7 +200,7 @@ AbstractLog::free(Reference reference)
     segment->trackDeadEntry(type, lengthWithMetadata);
     if (type == LOG_ENTRY_TYPE_OBJ)
         totalBytesRemaining += lengthWithMetadata;
-    //TODO(seojin): handle RpcRecord and PreparedOp.
+    //TODO(seojin): handle RpcResult and PreparedOp.
 }
 
 /**
@@ -381,7 +381,7 @@ AbstractLog::append(Lock& appendLock,
     head->trackNewEntry(type, lengthWithMetadata);
     if (type == LOG_ENTRY_TYPE_OBJ)
         totalBytesRemaining -= lengthWithMetadata;
-    //TODO(seojin): handle RpcRecord and PreparedOp.
+    //TODO(seojin): handle RpcResult and PreparedOp.
 
     metrics.totalBytesAppended += length;
     metrics.totalMetadataBytesAppended += (lengthWithMetadata - length);
