+++ src/LogEntryTypes.cc
@@ -40,7 +40,7 @@ toString(LogEntryType type)
         return "Object Safe Version";
     case LOG_ENTRY_TYPE_TABLESTATS:
         return "Table Stats Digest";
-    case LOG_ENTRY_TYPE_RPCRECORD:
+    case LOG_ENTRY_TYPE_RPCRESULT:
         return "Linearizable Rpc Record";
     case LOG_ENTRY_TYPE_PREP:
         return "Transaction Prepare Record";
