+++ src/LogEntryTypes.h
@@ -46,8 +46,8 @@ enum LogEntryType {
     /// See TableStats.h::Digest
     LOG_ENTRY_TYPE_TABLESTATS,
 
-    /// See Object.h::RpcRecord
-    LOG_ENTRY_TYPE_RPCRECORD,
+    /// See RpcResult.h
+    LOG_ENTRY_TYPE_RPCRESULT,
 
     /// See PreparedWrites.h::PreparedOp
     LOG_ENTRY_TYPE_PREP,
