+++ src/MakefragTest
@@ -115,6 +115,7 @@ TESTS_SRCFILES := \
 		  src/RecoveryTest.cc \
 		  src/ReplicaManagerTest.cc \
 		  src/ReplicatedSegmentTest.cc \
+		  src/RpcResultTest.cc \
 		  src/RpcTrackerTest.cc \
 		  src/RpcWrapperTest.cc \
 		  src/RuntimeOptionsTest.cc \
