+++ src/MasterService.h
@@ -27,6 +27,7 @@
 #include "ObjectFinder.h"
 #include "ObjectManager.h"
 #include "ReplicaManager.h"
+#include "RpcResult.h"
 #include "SegmentIterator.h"
 #include "SegmentManager.h"
 #include "ServerConfig.h"
@@ -322,7 +323,7 @@ class MasterService : public Service {
         Buffer resultBuffer;
         Log::Reference resultRef((uint64_t)result);
         objectManager.getLog()->getEntry(resultRef, resultBuffer);
-        RpcRecord savedRec(resultBuffer);
+        RpcResult savedRec(resultBuffer);
         return *((typename LinearizableRpcType::Response*) savedRec.getResp());
     }
 
@@ -337,7 +338,7 @@ class MasterService : public Service {
         Buffer resultBuffer;
         Log::Reference resultRef(result);
         objectManager.getLog()->getEntry(resultRef, resultBuffer);
-        RpcRecord savedRec(resultBuffer);
+        RpcResult savedRec(resultBuffer);
         return *((WireFormat::TxPrepare::Vote*) savedRec.getResp());
     }
