+++ src/Object.cc
@@ -1139,306 +1139,4 @@ ObjectSafeVersion::computeChecksum()
     crc.update(&header, downCast<uint32_t>(OFFSET_OF(Header, checksum)));
     return crc.getResult();
 }
-
-/**
- * Construct an RpcRecord in preparation for storing it in the log.
- * This form is used when the header information is available in
- * individual pieces, while the response is stored in a Buffer
- * (typical use: during linearizable write RPCs).
- *
- * \param tableId
- *      TableId for the object modified by this RPC.
- * \param keyHash
- *      KeyHash for the object modified by this RPC.
- * \param leaseId
- *      leaseId given for this linearizable RPC.
- * \param rpcId
- *      rpcId given for this linearizable RPC.
- * \param ackId
- *      ackId given for this linearizable RPC.
- * \param respBuffer
- *      Buffer containing all chunks that will comprise this RPC's
- *      response.
- * \param respOff
- *      Byte offset in the buffer where response start
- * \param respLen
- *      Length of response.
- *      A length of 0 means that the RpcRecord occupies the entire buffer
- *      starting at offset.
- */
-RpcRecord::RpcRecord(uint64_t tableId,
-                     KeyHash keyHash,
-                     uint64_t leaseId,
-                     uint64_t rpcId,
-                     uint64_t ackId,
-                     Buffer& respBuffer,
-                     uint32_t respOff,
-                     uint32_t respLen)
-    : header(tableId,
-             keyHash,
-             leaseId,
-             rpcId,
-             ackId),
-      respLength(),
-      response(),
-      respBuffer(&respBuffer),
-      respOffset(respOff)
-{
-    // compute the actual default value
-    if (respLen == 0)
-        respLength = this->respBuffer->size() - respOffset;
-    else
-        respLength = respLen;
-
-    //TODO(seojin): is that needed?
-    void* retPtr;
-    if (respBuffer.peek(respOffset, &retPtr) >= respLength)
-        response = static_cast<char*>(retPtr);
-}
-
-/**
- * Construct an RpcRecord in preparation for storing it in the log.
- * This form is used when the header information is available in
- * individual pieces, and memory for response is already prepared.
- * (typical use: during linearizable write RPCs).
- *
- * \param tableId
- *      TableId for the object modified by this RPC.
- * \param keyHash
- *      KeyHash for the object modified by this RPC.
- * \param leaseId
- *      leaseId given for this linearizable RPC.
- * \param rpcId
- *      rpcId given for this linearizable RPC.
- * \param ackId
- *      ackId given for this linearizable RPC.
- * \param response
- *      Buffer containing all chunks that will comprise this RPC's
- *      response.
- * \param respLen
- *      Length of response.
- */
-RpcRecord::RpcRecord(uint64_t tableId,
-                     KeyHash keyHash,
-                     uint64_t leaseId,
-                     uint64_t rpcId,
-                     uint64_t ackId,
-                     const void* response,
-                     uint32_t respLen)
-    : header(tableId,
-             keyHash,
-             leaseId,
-             rpcId,
-             ackId),
-      respLength(respLen),
-      response(response),
-      respBuffer(),
-      respOffset()
-{
-}
-
-/**
- * Construct a RpcRecord by deserializing an existing RpcRecord. Use
- * this constructor when reading existing RpcRecord from the log or from
- * individual log segments.
- *
- * \param buffer
- *      Buffer pointing to a complete serialized RpcRecord. It is the
- *      caller's responsibility to make sure that the buffer passed in
- *      actually contains a full RpcRecord. If it does not, then behavior
- *      is undefined.
- * \param offset
- *      Starting offset in the buffer where the RpcRecord begins.
- * \param length
- *      Total length of the RpcRecord in bytes.
- *      A length of 0 means that the RpcRecord occupies the entire buffer
- *      starting at offset.
- */
-RpcRecord::RpcRecord(Buffer& buffer, uint32_t offset, uint32_t length)
-    : header(*buffer.getOffset<Header>(offset)),
-      respLength(),
-      response(),
-      respBuffer(&buffer),
-      respOffset(sizeof32(header) + offset)
-{
-    if (length == 0)
-        respLength = downCast<uint16_t>(buffer.size() -
-                  offset - sizeof32(Header));
-    else
-        respLength = downCast<uint16_t>(length - sizeof32(Header));
-}
-
-/**
- * Append the full RpcRecord, including header, and response exactly as it
- * should be stored in the log, to a buffer
- *
- * \param buffer
- *      The buffer to append a serialized version of this RpcRecord to.
- */
-void
-RpcRecord::assembleForLog(Buffer& buffer)
-{
-    header.checksum = computeChecksum();
-    buffer.appendExternal(&header, sizeof32(header));
-    //buffer.appendExternal(&clusterTime, 8);
-    appendRespToBuffer(buffer);
-}
-
-/**
- * Copy the full serialized RpcRecord including the header and the response
- * to a contiguous region of memory memBlock.
- *
- * \param memBlock
- *      Destination to copy a serialized version of this RpcRecord to.
- *      The caller must ensure that sufficient memory is allocated for the
- *      complete RpcRecord to be stored.
- */
-void
-RpcRecord::assembleForLog(void* memBlock)
-{
-    uint8_t *dst = reinterpret_cast<uint8_t*>(memBlock);
-    header.checksum = computeChecksum();
-
-    memcpy(dst, &header, sizeof32(header));
-    //memcpy(dst + sizeof32(header), &clusterTime, 8);
-    memcpy(dst + sizeof32(header), getResp(), respLength);
-}
-
-/**
- * Append the response portion of this RpcRecord to a provided
- * buffer.
- *
- * \param buffer
- *      The buffer to append the response to.
- */
-void
-RpcRecord::appendRespToBuffer(Buffer& buffer)
-{
-    if (response) {
-        buffer.appendExternal(response, getRespLength());
-        return;
-    }
-
-    Buffer::Iterator it(respBuffer, respOffset, getRespLength());
-
-    while (!it.isDone()) {
-        buffer.appendExternal(it.getData(), it.getLength());
-        it.next();
-    }
-}
-
-/**
- * Obtain the 64-bit table identifier for the object modified by this RPC.
- */
-uint64_t
-RpcRecord::getTableId()
-{
-    return header.tableId;
-}
-
-/**
- * Obtain the KeyHash for the object modified by this RPC.
- */
-KeyHash
-RpcRecord::getKeyHash()
-{
-    return header.keyHash;
-}
-
-/**
- * Obtain the 64-bit lease identifier given for this linearizable RPC.
- */
-uint64_t
-RpcRecord::getLeaseId()
-{
-    return header.leaseId;
-}
-
-/**
- * Obtain the 64-bit RPC identifier given for this linearizable RPC.
- */
-uint64_t
-RpcRecord::getRpcId()
-{
-    return header.rpcId;
-}
-
-/**
- * Obtain the 64-bit acknowledgment value given for this linearizable RPC.
- */
-uint64_t
-RpcRecord::getAckId()
-{
-    return header.ackId;
-}
-
-/**
- * Obtain a pointer to a contiguous copy of this RPC's response.
- *
- * \param[out] length
- *      The length of the RPC's response in bytes.
- *
- * \return
- *      NULL if the RpcRecord is malformed,
- *      a pointer to a contiguous copy of the RPC's response otherwise
- */
-const void*
-RpcRecord::getResp(uint32_t *length)
-{
-    if (length)
-        *length = respLength;
-
-    if (response)
-        return static_cast<const uint8_t*>(response);
-    else
-        return respBuffer->getRange(respOffset, respLength);
-}
-
-uint32_t
-RpcRecord::getRespLength()
-{
-    return respLength;
-}
-
-/**
- * Compute a checksum on the RpcRecord and determine whether or not it matches
- * what is stored in the RpcRecord. Returns true if the checksum looks ok,
- * otherwise returns false.
- */
-bool
-RpcRecord::checkIntegrity()
-{
-    return computeChecksum() == header.checksum;
-}
-
-/**
- * Obtain the total size of the RpcRecord including its header.
- */
-uint32_t
-RpcRecord::getSerializedLength()
-{
-    return sizeof32(header) + respLength;
-}
-
-/**
- * Compute the RpcRecord's checksum and return it.
- */
-uint32_t
-RpcRecord::computeChecksum()
-{
-    assert(OFFSET_OF(Header, checksum) ==
-        (sizeof(header) - sizeof(header.checksum)));
-
-    Crc32C crc;
-    crc.update(&header, downCast<uint32_t>(OFFSET_OF(Header, checksum)));
-
-    if (response) {
-        crc.update(response, getRespLength());
-    } else {
-        crc.update(*respBuffer, respOffset, getRespLength());
-    }
-
-    return crc.getResult();
-}
-
 } // namespace RAMCloud
